Refactor: Split LegacyEntityDaos.kt into per-DAO files
Extract UserDao, CourseDao, CourseStepDao, ExamDao, QuestionDao,
SubmissionDao, AnswerDao, and TeamDao from LegacyEntityDaos.kt into
their own individual files in `org.ole.planet.myplanet.data.room.dao`.
This adheres to the repository's `*Dao.kt` naming convention and
simplifies future modifications to the data layer. No query text or
functional logic was altered.
