+++ app/src/main/java/org/ole/planet/myplanet/data/room/dao/CourseStepDao.kt
@@ -1,0 +1,15 @@
+package org.ole.planet.myplanet.data.room.dao
+
+import androidx.room.Dao
+import androidx.room.Query
+import androidx.room.Upsert
+import org.ole.planet.myplanet.model.CourseStep
+
+@Dao
+interface CourseStepDao {
+    @Query("SELECT * FROM course_steps WHERE courseId = :courseId") suspend fun getByCourseId(courseId: String): List<CourseStep>
+    @Query("SELECT * FROM course_steps WHERE courseId IN (:courseIds)") suspend fun getByCourseIds(courseIds: List<String>): List<CourseStep>
+    @Query("SELECT * FROM course_steps WHERE id = :id LIMIT 1") suspend fun getById(id: String): CourseStep?
+    @Upsert suspend fun upsertAll(items: List<CourseStep>)
+    @Upsert fun upsertAllBlocking(items: List<CourseStep>)
+}
